move alert to legacySharedComponents
